feat(money): store amounts as integer cents
Introduce Money(@JvmInline value class) backed by Long cents to eliminate
floating-point rounding. SQL schema changes REAL→INTEGER; a destructive
onUpgrade callback handles alpha-stage DB migration. All domain entities
(Transaction, TransactionInsert, TransactionUpdate, TransactionWithCategory,
HomeData) now carry Money. Mappers bridge Long↔Money. A new MoneyFormatter
helper in app/core/format/ handles locale-aware display. Tests migrated and
a regression test confirms three Money(33) values sum to exactly Money(99).

+++ data/src/main/kotlin/com/emm/data/account/AccountLocalDataSource.kt
@@ -61,7 +61,7 @@ class AccountLocalDataSource(private val emmDatabase: EmmDatabaseData) {
         )
     }
 
-    suspend fun getBalance(accountId: String): Double = withContext(Dispatchers.IO) {
+    suspend fun getBalance(accountId: String): Long = withContext(Dispatchers.IO) {
         emmDatabase.transactionsQueries.getAccountBalance(accountId).executeAsOne()
     }
 }
